backend-langgraphapi/app/agents/assistant0.py
from langgraph.prebuilt import ToolNode, create_react_agent
#from langchain_openai import ChatOpenAI
from langchain_google_genai import ChatGoogleGenerativeAI
from datetime import date
from app.agents.tools.user_info import get_user_info
from app.agents.tools.project_tasks import get_task_details

tools = [get_user_info, get_task_details]
#tools = [get_user_info]

#llm = ChatOpenAI(model="gpt-4.1-mini")

llm = ChatGoogleGenerativeAI(model="gemini-2.5-flash", project="key-airlock-477208-r8")

def get_prompt():
    today_str = date.today().strftime('%Y-%m-%d')
    return (
        f"You are a personal assistant named Assistant0. You are a helpful assistant that can answer questions and help with tasks. "
        f"Today's date is {today_str}. You have access to a set of tools, use the tools as needed to answer the user's question. "
        f"About get_details_task Tool, use current logged in user name as default assignee argument if not explicitely specified. Do not retun the task Id"
        f"Render the email body as a markdown block, do not wrap it in code blocks."
    )

agent = create_react_agent(
        model=llm,
        tools=ToolNode(tools, handle_tool_errors=False),
        prompt=get_prompt()
)
